fix: simplify
